+++ lib/hashie/dash.rb
@@ -1,5 +1,8 @@
 require 'hashie/hash'
 require 'set'
+require 'axiom-types'
+
+class Boolean; end
 
 module Hashie
   # A Dash is a 'defined' or 'discrete' Hash, that is, a Hash
@@ -47,19 +50,23 @@ module Hashie
         @subclasses.each { |klass| klass.property(property_name, options) }
       end
       required_properties << property_name if options.delete(:required)
-
-      enumerated_types[property_name.to_sym] = options.delete(:in) if options.key?(:in)
+      set_constraint(property_name, options.delete(:constraints)) if options.key?(:constraints)
     end
 
     class << self
       attr_reader :properties, :defaults
       attr_reader :required_properties
-      attr_reader :enumerated_types
+      attr_reader :property_constraints
+
+      def constraint_builders
+        @@constraint_builders
+      end
     end
     instance_variable_set('@properties', Set.new)
     instance_variable_set('@defaults', {})
     instance_variable_set('@required_properties', Set.new)
-    instance_variable_set('@enumerated_types', {})
+    instance_variable_set('@property_constraints', {})
+    class_variable_set('@@constraint_builders', {})
 
     def self.inherited(klass)
       super
@@ -67,7 +74,7 @@ module Hashie
       klass.instance_variable_set('@properties', properties.dup)
       klass.instance_variable_set('@defaults', defaults.dup)
       klass.instance_variable_set('@required_properties', required_properties.dup)
-      klass.instance_variable_set('@enumerated_types', enumerated_types.dup)
+      klass.instance_variable_set('@property_constraints', property_constraints.dup)
     end
 
     # Check to see if the specified property has already been
@@ -82,8 +89,85 @@ module Hashie
       required_properties.include? name
     end
 
-    def self.enumerated_type?(name)
-      enumerated_types.key?(name.to_sym)
+    # Check to see if the specified property is
+    # constrained.
+    def self.constrained?(name)
+      property_constraints.key?(name)
+    end
+
+    # Set the constraint for the specified property
+    def self.set_constraint(property_name, constraints)
+      raw_type           = constraints.key?(:type) ? constraints.delete(:type) : ::Object
+
+      assert_constraint_type_exists!(property_name, raw_type)
+
+      custom_constraints = build_custom_constraints(constraints)
+
+      begin
+        constraint_klass   = ::Axiom::Types.const_get(raw_type.to_s)
+      rescue NameError
+        fail_invalid_constraint_key!(property_name, :type)
+      end
+
+      constraint_type    = constraint_klass.new do
+        custom_constraints.each do |custom_constraint|
+          constraint(&custom_constraint)
+        end
+      end
+
+      # This is a bit of a hack
+      #
+      # Axiom allows you to create subtypes of types with constraints
+      # and the descendents inherit the constraints.
+      #
+      # When we build a new Axiom type with a block, if any of the constraints
+      # are invalid, we'll get an error but no information about which of
+      # the constraints caused it.
+      #
+      # Instead, we create a type hierarchy with each generation corresponding
+      # to one constraint, so we can identify it if it fails
+      # this is purely to give a dev-friendly exception message.
+      #
+      # It could be replaced with a more efficient method that only creates
+      # one type, if the ability to detect specific faulty constraints
+      # is not required.
+      klass = self
+      constraints.each_pair do |axiom_constraint, value|
+        begin
+          constraint_type = constraint_type.new do
+            send(axiom_constraint, value)
+          end
+        rescue NameError, ArgumentError
+          klass.fail_invalid_constraint_key!(property_name, axiom_constraint)
+        end
+      end
+
+      property_constraints[property_name] = constraint_type
+    end
+
+    # Register new custom constraint builders
+    # A constraint builder is a Proc that returns
+    # a new Proc bound with variables
+    #
+    # See underneath for examples
+    def self.register_constraint_builder(name, &block)
+      constraint_builders[name] = block
+    end
+
+    # Register the 'in' constraint, that ensures the value is
+    # included in the specified list
+    #
+    # class StateMachine < Hashie::Dash
+    #   property :current, constraints: { in: [:open,:close,:error] }
+    # end
+    #
+    register_constraint_builder(:in) do |list_of_values|
+      ->(value) { list_of_values.include?(value) }
+    end
+
+    # Register the exact length constraint
+    register_constraint_builder(:length) do |length|
+      ->(value) { value.respond_to?(:length) && value.length == length }
     end
 
     # You may initialize a Dash with an attributes hash
@@ -126,7 +210,7 @@ module Hashie
     def []=(property, value)
       assert_property_required! property, value
       assert_property_exists! property
-      assert_property_valid_enumerated_type! property, value
+      assert_property_within_constraints! property, value
       super(property, value)
     end
 
@@ -167,6 +251,21 @@ module Hashie
 
     private
 
+    def self.build_custom_constraints(constraints)
+      [].tap do |custom_constraints|
+        constraints.each_pair do |key, value|
+          # Is the constraint a pre-built custom constraint?
+          if value.is_a?(Proc)
+            custom_constraints << constraints.delete(key)
+
+          # Is the constraint a reference to a constraint builder?
+          elsif constraint_builders.key?(key)
+            custom_constraints << constraint_builders[key].call(constraints.delete(key))
+          end
+        end
+      end
+    end
+
     def initialize_attributes(attributes)
       attributes.each_pair do |att, value|
         self[att] = value
@@ -193,8 +292,19 @@ module Hashie
       fail_property_required_error!(property) if self.class.required?(property) && value.nil?
     end
 
-    def assert_property_valid_enumerated_type!(property, value)
-      fail_invalid_enumerated_type_error(property, value) if self.class.enumerated_type?(property) && value && !self.class.enumerated_types[property.to_sym].include?(value)
+    def assert_property_within_constraints!(property, value)
+      # RuboCop gets it wrong here, as we need to
+      # differentiate between nil and false
+      #
+      # (!value && value == false) would be the alternative, but
+      # it's very verbose compared to an explicit check
+      # rubocop:disable NonNilCheck
+      fail_property_outside_constraints!(property, value) if self.class.constrained?(property) && !value.nil? && !self.class.property_constraints[property].include?(value)
+      # rubocop:enable NonNilCheck
+    end
+
+    def self.assert_constraint_type_exists!(property, type)
+      fail_invalid_constraint_key!(property, :type) unless Axiom::Types.const_defined?(type.to_s) rescue false
     end
 
     def fail_property_required_error!(property)
@@ -205,8 +315,12 @@ module Hashie
       fail NoMethodError, "The property '#{property}' is not defined for #{self.class.name}."
     end
 
-    def fail_invalid_enumerated_type_error(property, value)
-      fail ArgumentError, "'#{value}' is not a valid value for the property '#{property}' for #{self.class.name}"
+    def fail_property_outside_constraints!(property, value)
+      fail ArgumentError, "The value '#{value}:#{value.class}' does not meet the constraints of the property '#{property}' for #{self.class.name}."
+    end
+
+    def self.fail_invalid_constraint_key!(property, key)
+      fail ArgumentError, "The constraint key '#{key}' is invalid for '#{property}' for #{name}."
     end
   end
 end